requirements.txt
streamlit==1.29.0
numpy==1.26.3
opencv-python==4.9.0.80
tensorflow==2.15.0
pillow==10.2.0
matplotlib==3.8.2